ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
pypdf
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-chroma
langchain-core
langchain-huggingface
chromadb
sentence-transformers